pyproject.toml
[tool.poetry]
name = "picasso-task"
version = "0.1.0"
description = ""
authors = ["suband74 <suband@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.3"
SQLAlchemy = "^1.4.39"
click = "^8.1.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
